requirements.txt
django
psycopg2-binary
dj-database-url
python-dotenv
pillow